+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 **GoSS**
 - Fix flag and argument ordering.
+- Add `package` parameter to appropriate methods.
+- Add `vars` parameter to `render` method.
 
 **Helm**
 - Add `atomic`, `version`, `devel`, and `createNS` parameters to appropriate methods.